js: Rename night_mode.ts to dark_theme.ts.
