+++ tools/test-js-with-node
@@ -59,6 +59,7 @@ EXEMPT_FILES = {
     "static/js/copy_and_paste.js",
     "static/js/csrf.ts",
     "static/js/css_variables.js",
+    "static/js/dark_theme.ts",
     "static/js/debug.js",
     "static/js/deprecated_feature_notice.js",
     "static/js/desktop_integration.js",
@@ -101,7 +102,6 @@ EXEMPT_FILES = {
     "static/js/muted_users_ui.js",
     "static/js/narrow.js",
     "static/js/navigate.js",
-    "static/js/night_mode.ts",
     "static/js/notifications.js",
     "static/js/overlays.js",
     "static/js/padded_widget.ts",
